Rename all evaluation methods to Query.EnvironmentalResult and merge into single file
